Added Alps PCB
Copied the kicad project to add an Alps version of the PCB
Switched to Kicad 6 for Alps version
Also generated the gerbers, BOPM and CPL
Sidenote : JLCPCB now accepts installation of USB-C port, but it shows it rotated 90° on the placement preview. I haven't ordered the project and most certainly will not, but it's something to look for.